ch: factor out poker_player and pot
 - factor out poker_player and pot

[Finishes #172825140]
